agent: Rename `request.py` to `models.py`
